+++ edison-jobs/src/main/java/de/otto/edison/jobs/eventbus/JobStateChangeListener.java
@@ -1,5 +1,6 @@
 package de.otto.edison.jobs.eventbus;
 
+import de.otto.edison.jobs.eventbus.events.MessageEvent;
 import de.otto.edison.jobs.eventbus.events.StateChangeEvent;
 import org.springframework.context.event.EventListener;
 
@@ -8,4 +9,6 @@ public interface JobStateChangeListener {
     @EventListener
     void consumeStateChange(StateChangeEvent stateChangeEvent);
 
+    @EventListener
+    void consumeMessage(MessageEvent messageEvent);
 }