+++ edison-jobs/src/main/java/de/otto/edison/jobs/eventbus/LogJobStateChangeListener.java
@@ -1,8 +1,13 @@
 package de.otto.edison.jobs.eventbus;
 
+import de.otto.edison.jobs.eventbus.events.MessageEvent;
 import de.otto.edison.jobs.eventbus.events.StateChangeEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import static de.otto.edison.jobs.domain.Level.ERROR;
+import static de.otto.edison.jobs.domain.Level.INFO;
+import static de.otto.edison.jobs.domain.Level.WARNING;
 
 public class LogJobStateChangeListener implements JobStateChangeListener {
 
@@ -16,6 +21,19 @@ public class LogJobStateChangeListener implements JobStateChangeListener {
                 stateChangeEvent.getState(),
                 stateChangeEvent.getJobId(),
                 stateChangeEvent.getMessage());
+    }
+
+    @Override
+    public void consumeMessage(final MessageEvent messageEvent) {
+        if (messageEvent.getLevel() == INFO) {
+            LOG.info("message='{}'", messageEvent.getMessage());
+        }
+        if (messageEvent.getLevel() == WARNING) {
+            LOG.warn("message='{}'", messageEvent.getMessage());
+        }
+        if (messageEvent.getLevel() == ERROR) {
+            LOG.error("message='{}'", messageEvent.getMessage());
+        }
     }
 
 }