+++ edison-jobs/src/main/java/de/otto/edison/jobs/service/JobRunnable.java
@@ -1,6 +1,7 @@
 package de.otto.edison.jobs.service;
 
 import de.otto.edison.jobs.definition.JobDefinition;
+import de.otto.edison.jobs.eventbus.JobEventPublisher;
 
 /**
  * The executable part of a Job that is executing background tasks in Edison microservices.
@@ -22,6 +23,20 @@ public interface JobRunnable {
      *
      * @return false, if the job was skipped without doing anything, true otherwise
      */
-    boolean execute();
+    default boolean execute() {
+        return true;
+    }
+
+    /**
+     * Executes the background task of the job and updates the JobInfo during execution.
+     * Inject JobEventPublisher to execute for backwards compatibility to branch 1.2.x
+     *
+     * @param jobEventPublisher publishes events to the event bus.
+     * @return false, if the job was skipped without doing anything, true otherwise
+     * @Deprecated use execute, don't remove this method
+     */
+    default boolean execute(JobEventPublisher jobEventPublisher) {
+        return execute();
+    }
 
 }